feat(ObserveViewport): reduce cpu usage on idle
The CPU usage on low level devices was in some situations up to 50% all the time because every
ObserveViewport component on the page did its own calculations whether to update or not.
The work performed by those components was always the same. To reduce the stress on the CPU
the Provider component now does all the required work and notifies its children as soon as
something changes.
To make this performant it was required to wrap the setState function of the ObserveViewport
components, which then rerenders children if required, into another requestAnimationFrame which
is considered an breaking change (even if I was not able to see any differences within the examples).
The onUpdate property is still fired on the first update and should anyhow be used in performance
heavy situations.