feat(ObserveViewport): implement deferUpdateUntilIdle option
deferUpdateUntilIdle allows to defer the update as long as no events
are collected anymore. Regardless whether the component is only
listening for e.g. dimensions updates, it is also defered for scroll
updates. This option is useful for updates that are optional and can
be defered until the page has time to do so. The onUpdate function
is always triggered with the latest viewport state.

At the moment the implmentation is based on a simple debounce
functionality with a wait of 700ms. In case it makes sense it
might get smarter in the future.