fix(ViewportCollector): detect resizes on the body element
Before this change e.g. enabling or disabling the scroll bars,
which change the documentHeight in most of the cases, was not
detected.

The fix requires a the faily new ResizeObserver API which is not
widley supported at the moment. Because this is an edge case that
might not be required for all applications we display a warning
and suggest to add a polyfill in case its an issue so everyone can
decide on their own if its worth it.