fix(ViewportProvider): allow to update cache for fallback once per second
It is actually possibe that the fallback needs to be accessed more than
once per session, e.g. when the listener count gets down to 0 and
increases afterwards again.
In this case the state should be recreated, but not for every component
that requests it.
The assumtion is, that this mostly happens on mount and the collector
should soon take over, therefore we define a limit of maximum one
request per second, to be up  to date but don't implement a performance
issue.

Fixes #10