feat(ViewportCollector): reduce work in idle mode
The raf is always ticking, even if no scroll or resize was triggered.
Now we initialize a raf if a scroll or resize element was performed if
none was regiestered earlier.
Also, this moves the layout triggering apis into the
requestAnimationFrame. This should not have big impact as browser are
optimizing this already (as far as I can read from the dev tools), but
we safe some unneeded JS in the handlers for merging objects.